+++ src/mattermostautodriver/endpoints/permissions.py
@@ -3,13 +3,9 @@ from .base import Base
 
 class Permissions(Base):
 
-    def get_ancillary_permissions(self, params=None):
+    def get_ancillary_permissions_post(self, options):
         """Return all system console subsection ancillary permissions
-
-        subsection_permissions: The subsection permissions to return the ancillary permissions for. These values are comma seperated. Ex. subsection_permissions=sysconsole_read_reporting_site_statistics,sysconsole_write_reporting_site_statistics,sysconsole_write_user_management_channels
-
-
-        `Read in Mattermost API docs (permissions - GetAncillaryPermissions) <https://api.mattermost.com/#tag/permissions/operation/GetAncillaryPermissions>`_
+        `Read in Mattermost API docs (permissions - GetAncillaryPermissionsPost) <https://api.mattermost.com/#tag/permissions/operation/GetAncillaryPermissionsPost>`_
 
         """
-        return self.client.get("""/api/v4/permissions/ancillary""", params=params)
+        return self.client.post("""/api/v4/permissions/ancillary""", options=options)